refactor directory strucutre
