+++ src/main/java/frc/robot/subsystems/swerve/SwerveModuleNeoTurnNeoDrive.java
@@ -14,9 +14,9 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
-import frc.robot.Constants;
-import frc.robot.Conversions;
-import frc.robot.RevOptimizer;
+import frc.robot.util.Constants;
+import frc.robot.util.Conversions;
+import frc.robot.util.RevOptimizer;
 
 /** 
  * Represents a swerve module with a NEO (SparkMAX) turn motor and a NEO (SparkMAX) drive motor.
