+++ src/main/java/frc/robot/util/Constants.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot;
+package frc.robot.util;
 
 import java.util.Map;
 import java.util.HashMap;
@@ -18,6 +18,7 @@ import edu.wpi.first.wpilibj.SPI;
 import com.pathplanner.lib.util.ReplanningConfig;
 import com.pathplanner.lib.util.HolonomicPathFollowerConfig;
 import com.pathplanner.lib.util.PIDConstants;
+import frc.robot.Robot;
 
 /**
  * The Constants class provides a convenient place for teams to hold robot-wide numerical or boolean
