+++ src/main/java/frc/robot/util/Conversions.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.util;
 
 public class Conversions {
